Update dorado_stereo.sh
When running with the -n flag the code in it's current form will attempt to make all the directories even though the command line argument suggests no commands will be run. 

This change should prevent that from happening.

The additional problem is that if the folders do not exist for the output, dorado won't create them - but that is obviously an issue for the base caller.